+++ .storybook/utils.tsx
@@ -1,8 +1,37 @@
-import { JSX, FC } from "react";
+import { JSX, FC, useInsertionEffect } from "react";
+import { MotifProvider } from "../src/lib";
+import { RESET_THEME_BUTTON_VAL } from "./preview.tsx";
 
-export const iconOptions = {
-  undefined: undefined,
-  string: "motif_ui",
+export const DEFAULT_THEME = "default-theme";
+
+const ThemeWrapper: FC<{ story: FC; theme: string }> = ({ story: Story, theme }) => {
+  useInsertionEffect(() => {
+    document.querySelectorAll("link[data-theme-css]").forEach(link => link.remove());
+
+    const link = document.createElement("link");
+    link.rel = "stylesheet";
+    link.setAttribute("data-theme-css", "true");
+    link.href = `/themes/${theme}.css`;
+    document.head.appendChild(link);
+
+    return () => link.remove();
+  }, [theme]);
+
+  return (
+    <MotifProvider>
+      <Story />
+    </MotifProvider>
+  );
+};
+
+export const themeChangeDecorator = (Story: FC, context: { globals: { theme?: string } }): JSX.Element => {
+  const themeProp = context.globals.theme as string;
+  const theme = !themeProp || themeProp === RESET_THEME_BUTTON_VAL ? DEFAULT_THEME : themeProp;
+  return <ThemeWrapper story={Story} theme={theme} />;
+};
+
+export const iconObjects = {
+  string: "folder",
   "<i>": <i className="bi bi-airplane" />,
   "<span>": <span className="material-symbols-outlined">thumb_up</span>,
   "<svg>": (
@@ -10,6 +39,11 @@ export const iconOptions = {
       <path d="M8 1.5 3 4v4c0 3 2 5 5 6 3-1 5-3 5-6V4z" />
     </svg>
   ),
+};
+
+export const iconOptions = {
+  undefined: undefined,
+  ...iconObjects,
 } as const;
 
 export const iconDecorator = (StoryComponent: FC): JSX.Element => (
