+++ src/lib/components/InputPassword/InputPassword.stories.tsx
@@ -8,6 +8,7 @@ const meta: Meta<typeof InputPassword> = {
   argTypes: {
     value: { type: "string" },
     size: { table: { defaultValue: { summary: "md" } } },
+    icon: { table: { defaultValue: { summary: "vpn_key" } } },
   },
   args: {
     placeholder: "Placeholder",
