+++ src/lib/components/ListView/ListView.module.scss
@@ -38,6 +38,7 @@ $list-item-sizes: (("sm", "xs", "2x", "16x", "8x", "24x", "20x", "22x"), ("md", 
 
 .Root {
   padding: 0;
+  margin: 0;
   list-style: none;
 }
